requirements.txt
certifi==2019.6.16
colored==1.3.93
packaging==19.2
Pillow==6.1.0
pip-review==1.0
pipenv==2018.11.26
pygame==1.9.4
pyparsing==2.4.2
six==1.12.0
termcolor==1.1.0
virtualenv==16.7.4
virtualenv-clone==0.5.3
